candy_store_server/candy_store/serializers.py
from rest_framework import serializers

from candy_store.models import Category


class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = ['category_id', 'category_name', 'slug', ]
