candy_store_server/pyproject.toml
[tool.poetry]
name = "candy-store-server"
version = "0.1.0"
description = ""
authors = ["Angelina_RLY <cher.rly2036@gmail.com>"]
readme = "README.md"
packages = [{include = "candy_store_server"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
redis = "^5.0.3"
hiredis = "^2.3.2"
pillow = "^10.2.0"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
flake8 = "^7.0.0"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
